ML-Pipeline/requirements.txt
# Combined Requirements File
# Install with: pip install -r requirements.txt

# Deep Learning Frameworks
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
tensorflow>=2.13.0

# Data Processing & Core Libraries
numpy>=1.24.0
pandas>=2.0.0
pillow>=10.0.0
opencv-python>=4.8.0

# Machine Learning
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0

# Utilities
tqdm>=4.62.0

transformers==4.35.0
